Refactor Requester class into schemas.py
